+++ tests/fixtures/modules/terraform-azure-avm-res-mock/examples/default/main.tf
@@ -3,6 +3,7 @@ module "test" {
 
   location                 = "westus3"
   create_example_resources = false
+  enable_telemetry         = false
 }
 
 output "resource_ids" {